reorganize deprecations
