+++ Sources/_SwiftUINavigationState/AlertState.swift
@@ -258,110 +258,3 @@ extension Alert {
     }
   }
 }
-
-// MARK: - Deprecations
-
-extension AlertState {
-  @available(*, deprecated, message: "Use 'ButtonState<Action>' instead.")
-  public typealias Button = ButtonState<Action>
-
-  @available(*, deprecated, message: "Use 'ButtonState<Action>.ButtonAction' instead.")
-  public typealias ButtonAction = ButtonState<Action>.ButtonAction
-
-  @available(*, deprecated, message: "Use 'ButtonState<Action>.Role' instead.")
-  public typealias ButtonRole = ButtonState<Action>.Role
-
-  @available(
-    iOS, introduced: 15, deprecated: 100000, message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    macOS,
-    introduced: 12,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    tvOS, introduced: 15, deprecated: 100000, message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    watchOS,
-    introduced: 8,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  public init(
-    title: TextState,
-    message: TextState? = nil,
-    buttons: [ButtonState<Action>]
-  ) {
-    self.init(
-      id: UUID(),
-      buttons: buttons,
-      message: message,
-      title: title
-    )
-  }
-
-  @available(
-    iOS, introduced: 13, deprecated: 100000, message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    macOS,
-    introduced: 10.15,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    tvOS, introduced: 13, deprecated: 100000, message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    watchOS,
-    introduced: 6,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  public init(
-    title: TextState,
-    message: TextState? = nil,
-    dismissButton: ButtonState<Action>? = nil
-  ) {
-    self.init(
-      id: UUID(),
-      buttons: dismissButton.map { [$0] } ?? [],
-      message: message,
-      title: title
-    )
-  }
-
-  @available(
-    iOS, introduced: 13, deprecated: 100000, message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    macOS,
-    introduced: 10.15,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    tvOS, introduced: 13, deprecated: 100000, message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    watchOS,
-    introduced: 6,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  public init(
-    title: TextState,
-    message: TextState? = nil,
-    primaryButton: ButtonState<Action>,
-    secondaryButton: ButtonState<Action>
-  ) {
-    self.init(
-      id: UUID(),
-      buttons: [primaryButton, secondaryButton],
-      message: message,
-      title: title
-    )
-  }
-}
