+++ Sources/_SwiftUINavigationState/ButtonState.swift
@@ -315,70 +315,6 @@ extension Button where Label == Text {
   }
 }
 
-// MARK: - Deprecations
-
-extension ButtonState {
-  @available(*, deprecated, renamed: "ButtonStateAction")
-  public typealias ButtonAction = ButtonStateAction<Action>
-
-  @available(*, deprecated, renamed: "ButtonStateRole")
-  public typealias Role = ButtonStateRole
-}
-
-extension ButtonStateAction {
-  @available(*, deprecated, message: "Use 'ButtonState.withAction' instead.")
-  public typealias ActionType = _ActionType
-}
-
-@available(
-  iOS,
-  introduced: 13,
-  deprecated: 100000,
-  message: "Use 'ButtonState.init(role:action:label:)' instead."
-)
-@available(
-  macOS, introduced: 10.15,
-  deprecated: 100000,
-  message: "Use 'ButtonState.init(role:action:label:)' instead."
-)
-@available(
-  tvOS,
-  introduced: 13,
-  deprecated: 100000,
-  message: "Use 'ButtonState.init(role:action:label:)' instead."
-)
-@available(
-  watchOS,
-  introduced: 6,
-  deprecated: 100000,
-  message: "Use 'ButtonState.init(role:action:label:)' instead."
-)
-extension ButtonState {
-  public static func cancel(
-    _ label: TextState, action: ButtonStateAction<Action> = .send(nil)
-  ) -> Self {
-    Self(role: .cancel, action: action) {
-      label
-    }
-  }
-
-  public static func `default`(
-    _ label: TextState, action: ButtonStateAction<Action> = .send(nil)
-  ) -> Self {
-    Self(action: action) {
-      label
-    }
-  }
-
-  public static func destructive(
-    _ label: TextState, action: ButtonStateAction<Action> = .send(nil)
-  ) -> Self {
-    Self(role: .destructive, action: action) {
-      label
-    }
-  }
-}
-
 @usableFromInline
 func debugCaseOutput(_ value: Any) -> String {
   func debugCaseOutputHelp(_ value: Any) -> String {
