+++ Sources/_SwiftUINavigationState/ConfirmationDialogState.swift
@@ -280,139 +280,3 @@ extension Visibility {
     }
   }
 }
-
-// MARK: - Deprecations
-
-@available(iOS 13, *)
-@available(macOS 12, *)
-@available(tvOS 13, *)
-@available(watchOS 6, *)
-extension ConfirmationDialogState {
-  @available(*, deprecated, message: "Use 'ButtonState<Action>' instead.")
-  public typealias Button = ButtonState<Action>
-
-  @available(*, deprecated, renamed: "ConfirmationDialogStateTitleVisibility")
-  public typealias Visibility = ConfirmationDialogStateTitleVisibility
-
-  @available(
-    iOS,
-    introduced: 13,
-    deprecated: 100000,
-    message: "Use 'init(titleVisibility:title:actions:message:)' instead."
-  )
-  @available(
-    macOS,
-    introduced: 12,
-    deprecated: 100000,
-    message: "Use 'init(titleVisibility:title:actions:message:)' instead."
-  )
-  @available(
-    tvOS,
-    introduced: 13,
-    deprecated: 100000,
-    message: "Use 'init(titleVisibility:title:actions:message:)' instead."
-  )
-  @available(
-    watchOS,
-    introduced: 6,
-    deprecated: 100000,
-    message: "Use 'init(titleVisibility:title:actions:message:)' instead."
-  )
-  public init(
-    title: TextState,
-    titleVisibility: ConfirmationDialogStateTitleVisibility,
-    message: TextState? = nil,
-    buttons: [ButtonState<Action>] = []
-  ) {
-    self.init(
-      id: UUID(),
-      buttons: buttons,
-      message: message,
-      title: title,
-      titleVisibility: titleVisibility
-    )
-  }
-
-  @available(
-    iOS,
-    introduced: 13,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    macOS,
-    introduced: 12,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    tvOS,
-    introduced: 13,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  @available(
-    watchOS,
-    introduced: 6,
-    deprecated: 100000,
-    message: "Use 'init(title:actions:message:)' instead."
-  )
-  public init(
-    title: TextState,
-    message: TextState? = nil,
-    buttons: [ButtonState<Action>] = []
-  ) {
-    self.init(
-      id: UUID(),
-      buttons: buttons,
-      message: message,
-      title: title,
-      titleVisibility: .automatic
-    )
-  }
-}
-
-@available(iOS, introduced: 13, deprecated: 100000, renamed: "ConfirmationDialogState")
-@available(macOS, introduced: 12, unavailable)
-@available(tvOS, introduced: 13, deprecated: 100000, renamed: "ConfirmationDialogState")
-@available(watchOS, introduced: 6, deprecated: 100000, renamed: "ConfirmationDialogState")
-public typealias ActionSheetState<Action> = ConfirmationDialogState<Action>
-
-@available(
-  iOS,
-  introduced: 13,
-  deprecated: 100000,
-  message:
-    "use 'View.confirmationDialog(title:isPresented:titleVisibility:presenting::actions:)' instead."
-)
-@available(
-  macOS,
-  introduced: 12,
-  unavailable
-)
-@available(
-  tvOS,
-  introduced: 13,
-  deprecated: 100000,
-  message:
-    "use 'View.confirmationDialog(title:isPresented:titleVisibility:presenting::actions:)' instead."
-)
-@available(
-  watchOS,
-  introduced: 6,
-  deprecated: 100000,
-  message:
-    "use 'View.confirmationDialog(title:isPresented:titleVisibility:presenting::actions:)' instead."
-)
-extension ActionSheet {
-  public init<Action>(
-    _ state: ConfirmationDialogState<Action>,
-    action: @escaping (Action?) -> Void
-  ) {
-    self.init(
-      title: Text(state.title),
-      message: state.message.map { Text($0) },
-      buttons: state.buttons.map { .init($0, action: action) }
-    )
-  }
-}
